autooooo

+++ sim_p1_v2/parallel/device_initialization.cu
@@ -21,7 +21,7 @@ __device__ device_cavi_implementation::device_cavi_implementation(device_setting
 
   // initialize estimates
   for(int k = 0; k < K; k++){
-    m[k] = curand_normal_double(&state) * sqrt(sigma_2);
+    m[k] = curand_normal_double(&state) * sqrt(double(sigma_2));
     s2[k] = sigma_2;
     for(int i = 0; i < n_samples; i++){
       phi[i][k] = 1/double(K);
